add: basic exceptions

+++ src/main/java/dev/anderson/livraria/service/LivrariaService.java
@@ -27,15 +27,25 @@ public class LivrariaService {
   }
 
   public LivroEntity getBookByNameAndYear(String name, Integer year) {
+    if (!livrariaRepository.existsByNameAndYear(name, year)) {
+      throw new DuplicatedBookException(HttpStatus.BAD_REQUEST, "Livro não encontrado");
+    }
     return livrariaRepository.findByNameAndYear(name, year);
   }
 
   public void deleteBook(Long id) {
+    if (!livrariaRepository.existsById(id)) {
+      throw new DuplicatedBookException(HttpStatus.BAD_REQUEST, "Livro não encontrado");
+    }
     livrariaRepository.deleteById(id);
   }
 
   public LivroEntity updateBook(Long id, LivroEntity livro) {
+    if (!livrariaRepository.existsById(id)) {
+      throw new DuplicatedBookException(HttpStatus.BAD_REQUEST, "Livro não encontrado");
+    }
     LivroEntity livroEntity = livrariaRepository.findById(id).get();
+
     livroEntity.update(livro);
     return livrariaRepository.save(livroEntity);
   }
